Changed memcmp, and made most functions recieve and give explicit sized types.

+++ include/xD-DOS/pmm.h
@@ -6,7 +6,7 @@
 
 #define PAGE_SIZE 4096
 
-int pmm_init();
+uint8_t pmm_init();
 void* pmm_alloc_page();
 void pmm_free_page(void *page);
